tests are done for functionality. Next the http will be added.

+++ cmpln/DBUpdate.go
@@ -5,7 +5,7 @@ import(
 )
 
 
-func UpdatePost(idnum uint, n, desc, topic string) (error) {
+func UpdatePost(idnum int64, n, desc, topic string) (error) {
     if err :=  SetupDBConn("root", "admin", "cmplnDB"); err != nil {
         return fmt.Errorf("Error trying to establish DB connection in UpdatePost-Function: %v", err)
     }
